gather utils for examples, fix rename metadata sources
