+++ examples/basic-io/idx_read.c
@@ -67,24 +67,18 @@
   #include "utils/PIDX_windows_utils.h"
 #endif
 
-enum { X, Y, Z, NUM_DIMS };
-static int process_count = 1, rank = 0;
-static unsigned long long local_box_offset[3];
-static unsigned long long global_box_size[3] = {0, 0, 0};
-static unsigned long long local_box_size[3] = {0, 0, 0};
-static int current_ts = 1;
-static int variable_index = 0;
-static char output_file_template[512] = "test_idx";
-static unsigned char *data;
-static PIDX_point global_size, local_offset, local_size, global_bounds;
-static PIDX_access p_access;
-static int variable_count;
-static PIDX_file file;
-static PIDX_variable variable;
-static int bits_per_sample = 0;
-static int values_per_sample = 0;
-static char type_name[512];
-static char output_file_name[512] = "test.idx";
+#include "pidx_examples_utils.h"
+
+static char input_file_name[512];
+int current_ts = 0;
+int variable_index = 0;
+int values_per_sample = 0;
+int bits_per_sample = 0;
+PIDX_variable variable;
+char type_name[512];
+static PIDX_point global_bounds;
+unsigned char *data;
+
 static char *usage = "Serial Usage: ./idx_read -g 32x32x32 -l 32x32x32 -v 0 -f input_idx_file_name\n"
                      "Parallel Usage: mpirun -n 8 ./idx_read -g 32x32x32 -l 16x16x16 -f -v 0 input_idx_file_name\n"
                      "  -g: global dimensions\n"
@@ -93,82 +87,60 @@ static char *usage = "Serial Usage: ./idx_read -g 32x32x32 -l 32x32x32 -v 0 -f i
                      "  -t: time step index to read\n"
                      "  -v: variable index to read";
 
-
-static void init_mpi(int argc, char **argv);
 static void parse_args(int argc, char **argv);
-static void check_args();
-static void calculate_per_process_offsets();
-static void terminate_with_error_msg(const char *format, ...);
-static void terminate();
-static void create_pidx_var_point_and_access();
-static void set_pidx_file(int ts);
 static void set_pidx_variable_and_create_buffer();
 static int verify_read_results();
-static void shutdown_mpi();
+static void set_pidx_file(int ts);
+
 
 int main(int argc, char **argv)
 {
   // Init MPI and MPI vars (e.g. rank and process_count)
   init_mpi(argc, argv);
-
+  
   // Parse input arguments and initialize
   // corresponing variables
   parse_args(argc, argv);
-
+  
   // Verify that the domain decomposition is valid
   // for the given number of cores
   check_args();
-
+  
   // Initialize per-process local domain
   calculate_per_process_offsets();
-
+  
   // Create variables
-  create_pidx_var_point_and_access();
-
+  create_pidx_point_and_access();
+  
   // Set PIDX_file for this timestep
   set_pidx_file(current_ts);
-
+  
   // Get all the information about the variable that we want to read
   set_pidx_variable_and_create_buffer();
-
+  
   // Read the data into a local buffer (data) in row major order
-  //printf("[INIT] %d ---- %lld %lld %lld - %lld %lld %lld\n", rank, local_offset[0], local_offset[1], local_offset[2], local_size[0], local_size[1], local_size[2]);
   PIDX_variable_read_data_layout(variable, local_offset, local_size, data, PIDX_row_major);
-
+  
   // PIDX_close triggers the actual write on the disk
   // of the variables that we just set
   PIDX_close(file);
-
+  
   // Close PIDX_access
   PIDX_close_access(p_access);
-
+  
   // Compare the data that we just against the syntethic data
   verify_read_results();
-
+  
   free(data);
   shutdown_mpi();
 }
-
-//----------------------------------------------------------------
-static void init_mpi(int argc, char **argv)
-{
-#if PIDX_HAVE_MPI
-  if (MPI_Init(&argc, &argv) != MPI_SUCCESS)
-    terminate_with_error_msg("ERROR: MPI_Init error\n");
-  if (MPI_Comm_size(MPI_COMM_WORLD, &process_count) != MPI_SUCCESS)
-    terminate_with_error_msg("ERROR: MPI_Comm_size error\n");
-  if (MPI_Comm_rank(MPI_COMM_WORLD, &rank) != MPI_SUCCESS)
-    terminate_with_error_msg("ERROR: MPI_Comm_rank error\n");
-#endif
-}
-
-
 
 static void parse_args(int argc, char **argv)
 {
   char flags[] = "g:l:f:t:v:";
   int one_opt = 0;
-
+  char input_file_template[512];
+  
   while ((one_opt = getopt(argc, argv, flags)) != EOF)
   {
     /* postpone error checking for after while loop */
@@ -187,9 +159,9 @@ static void parse_args(int argc, char **argv)
       break;
 
     case('f'): // input file name
-      if (sprintf(output_file_template, "%s", optarg) < 0)
+      if (sprintf(input_file_template, "%s", optarg) < 0)
         terminate_with_error_msg("Invalid output file name template\n%s", usage);
-      sprintf(output_file_name, "%s%s", output_file_template, ".idx");
+      sprintf(input_file_name, "%s%s", input_file_template, ".idx");
       break;
 
     case('t'): // number of timesteps
@@ -255,109 +227,44 @@ static void parse_args(int argc, char **argv)
 
 }
 
-static void check_args()
-{
-  if (global_box_size[X] < local_box_size[X] || global_box_size[Y] < local_box_size[Y] || global_box_size[Z] < local_box_size[Z])
-    terminate_with_error_msg("ERROR: Global box is smaller than local box in one of the dimensions\n");
-
-  // check if the number of processes given by the user is consistent with the actual number of processes needed
-  int brick_count = (int)((global_box_size[X] + local_box_size[X] - 1) / local_box_size[X]) *
-                    (int)((global_box_size[Y] + local_box_size[Y] - 1) / local_box_size[Y]) *
-                    (int)((global_box_size[Z] + local_box_size[Z] - 1) / local_box_size[Z]);
-  if(brick_count != process_count)
-    terminate_with_error_msg("ERROR: Number of sub-blocks (%d) doesn't match number of processes (%d)\n", brick_count, process_count);
-
-}
-
-static void calculate_per_process_offsets()
-{
-  int sub_div[NUM_DIMS];
-  sub_div[X] = (global_box_size[X] / local_box_size[X]);
-  sub_div[Y] = (global_box_size[Y] / local_box_size[Y]);
-  sub_div[Z] = (global_box_size[Z] / local_box_size[Z]);
-  local_box_offset[Z] = (rank / (sub_div[X] * sub_div[Y])) * local_box_size[Z];
-  int slice = rank % (sub_div[X] * sub_div[Y]);
-  local_box_offset[Y] = (slice / sub_div[X]) * local_box_size[Y];
-  local_box_offset[X] = (slice % sub_div[X]) * local_box_size[X];
-}
-
-//----------------------------------------------------------------
-static void terminate()
-{
-#if PIDX_HAVE_MPI
-  MPI_Abort(MPI_COMM_WORLD, -1);
-#else
-  exit(-1);
-#endif
-}
-
-//----------------------------------------------------------------
-static void terminate_with_error_msg(const char *format, ...)
-{
-  va_list arg_ptr;
-  va_start(arg_ptr, format);
-  vfprintf(stderr, format, arg_ptr);
-  va_end(arg_ptr);
-  terminate();
-}
-
-//----------------------------------------------------------------
-static void create_pidx_var_point_and_access()
-{
-  PIDX_set_point(global_size, global_box_size[X], global_box_size[Y], global_box_size[Z]);
-  PIDX_set_point(local_offset, local_box_offset[X], local_box_offset[Y], local_box_offset[Z]);
-  PIDX_set_point(local_size, local_box_size[X], local_box_size[Y], local_box_size[Z]);
-
-  //  Creating access
-  PIDX_create_access(&p_access);
-  PIDX_set_mpi_access(p_access, MPI_COMM_WORLD);
-
-  return;
-}
-
 //----------------------------------------------------------------
 static void set_pidx_file(int ts)
 {
   PIDX_return_code ret;
-
+  
   // Open IDX file
-  ret = PIDX_file_open(output_file_name, PIDX_MODE_RDONLY, p_access, global_bounds, &file);
+  ret = PIDX_file_open(input_file_name, PIDX_MODE_RDONLY, p_access, global_bounds, &file);
   if (ret != PIDX_success)  terminate_with_error_msg("PIDX_file_open\n");
-
+  
   // Set the current timestep
   PIDX_set_current_time_step(file, ts);
   // Get the total number of variables
   PIDX_get_variable_count(file, &variable_count);
-
-  PIDX_query_box(file, global_size);
-
-  //PIDX_disable_agg(file);
-  return;
 }
 
 //----------------------------------------------------------------
 static void set_pidx_variable_and_create_buffer()
 {
   PIDX_return_code ret;
-
+  
   // Check if the index variable requested is valid (< num variables in the dataset)
   if (variable_index >= variable_count) terminate_with_error_msg("Variable index more than variable count\n");
-
+  
   // Set the variable index that we want to read
   ret = PIDX_set_current_variable_index(file, variable_index);
   //ret = PIDX_set_current_variable_by_name(file, "var_1");
   if (ret != PIDX_success)  terminate_with_error_msg("PIDX_set_current_variable_index\n");
-
+  
   // Get corresponding PIDX_variable
   PIDX_get_current_variable(file, &variable);
-
+  
   // Get some information about this variable (typename, number of values per sample,
   // number of bits per sample)
   PIDX_values_per_datatype(variable->type_name, &values_per_sample, &bits_per_sample);
   strcpy(type_name, variable->type_name);
-
-  data = malloc((uint64_t)(bits_per_sample/8) * local_box_size[0] * local_box_size[1] * local_box_size[2]  * values_per_sample);
-  memset(data, 0, (uint64_t)(bits_per_sample/8) * local_box_size[0] * local_box_size[1] * local_box_size[2]  * values_per_sample);
+  
+  data = malloc((bits_per_sample/8) * local_box_size[0] * local_box_size[1] * local_box_size[2]  * values_per_sample);
+  memset(data, 0, (bits_per_sample/8) * local_box_size[0] * local_box_size[1] * local_box_size[2]  * values_per_sample);
 }
 
 //----------------------------------------------------------------
@@ -366,143 +273,89 @@ static int verify_read_results()
   int read_error_count = 0, read_count = 0;
   int total_read_error_count = 0, total_read_count = 0;
   int i, j, k, vps;
-  unsigned char uchar_val = 0;
-  short short_val = 0;
   int int_val = 0;
   double double_val = 0;
   float float_val = 0;
   int var = variable_index;
-
-  //if (rank == 100)
-  //    fprintf(stderr, "LBO %d %d %d LBS %d %d %d\n", local_box_offset[0], local_box_offset[1], local_box_offset[2], local_box_size[0], local_box_size[1], local_box_size[2]);
-
+  
   bits_per_sample = bits_per_sample / 8;
   for (k = 0; k < local_box_size[2]; k++)
     for (j = 0; j < local_box_size[1]; j++)
       for (i = 0; i < local_box_size[0]; i++)
       {
         unsigned long long index = (unsigned long long) (local_box_size[0] * local_box_size[1] * k) + (local_box_size[0] * j) + i;
-
-        if (strcmp(type_name, UINT8) == 0 || strcmp(type_name, UINT8_GA) == 0 || strcmp(type_name, UINT8_RGB) == 0)
-        {
-          for (vps = 0; vps < values_per_sample; vps++)
-          {
-            memcpy(&uchar_val, data + (index * values_per_sample + vps) * bits_per_sample, bits_per_sample);
-            if (uchar_val != var + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)))
-            {
-              read_error_count++;
-              //if (rank == 0)
-              //fprintf(stderr, "W[%d %d %d] [%d] Read error %d\n", i,j ,k, vps, int_val);//, var + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
-            }
-            else
-            {
-              read_count++;
-              //if (rank == 0)
-              //  fprintf(stderr, "C[%d %d %d] [%d] Read %f %lld\n", i,j ,k, vps, data[index * vps[var] + vps], var + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
-            }
-          }
-        }
-
-        else if (strcmp(type_name, INT16) == 0 || strcmp(type_name, INT16_GA) == 0 || strcmp(type_name, INT16_RGB) == 0)
-        {
-          for (vps = 0; vps < values_per_sample; vps++)
-          {
-            memcpy(&short_val, data + (index * values_per_sample + vps) * bits_per_sample, bits_per_sample);
-            if (short_val != var + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)))
-            {
-              read_error_count++;
-              //if (rank == 0)
-              //fprintf(stderr, "W[%d %d %d] [%d] Read error %d\n", i,j ,k, vps, int_val);//, var + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
-            }
-            else
-            {
-              read_count++;
-              //if (rank == 0)
-              //  fprintf(stderr, "C[%d %d %d] [%d] Read %f %lld\n", i,j ,k, vps, data[index * vps[var] + vps], var + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
-            }
-          }
-        }
-
-        else if (strcmp(type_name, INT32) == 0 || strcmp(type_name, INT32_GA) == 0 || strcmp(type_name, INT32_RGB) == 0)
+        
+        if (strcmp(type_name, INT32) == 0 || strcmp(type_name, INT32_GA) == 0 || strcmp(type_name, INT32_RGB) == 0)
         {
           for (vps = 0; vps < values_per_sample; vps++)
           {
             memcpy(&int_val, data + (index * values_per_sample + vps) * bits_per_sample, bits_per_sample);
-            if (int_val != var + 100 + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)))
+            if (int_val != var + 100 + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)))
             {
               read_error_count++;
               //if (rank == 0)
-              //fprintf(stderr, "W[%d %d %d] [%d] Read error %d\n", i,j ,k, vps, int_val);//, var + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
+              //  printf("W[%d %d %d] [%d] Read error %d\n", i,j ,k, vps, int_val);//, var + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
             }
             else
             {
               read_count++;
               //if (rank == 0)
-              //  fprintf(stderr, "C[%d %d %d] [%d] Read %f %lld\n", i,j ,k, vps, data[index * vps[var] + vps], var + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
+              //  printf("C[%d %d %d] [%d] Read %f %lld\n", i,j ,k, vps, data[index * vps[var] + vps], var + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
             }
           }
         }
-
+        
         else if (strcmp(type_name, FLOAT64) == 0 || strcmp(type_name, FLOAT64_RGB) == 0 || strcmp(type_name, FLOAT64_GA) == 0)
         {
           for (vps = 0; vps < values_per_sample; vps++)
           {
             memcpy(&double_val, data + (index * values_per_sample + vps) * bits_per_sample, bits_per_sample);
-            if (double_val != var + 100 + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)))
+            if (double_val != var + 100 + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)))
             {
               read_error_count++;
               //if (rank == 0)
-              //  fprintf(stderr, "W[%d %d %d] [%d] Read error %f %d\n", i,j ,k, vps, double_val,100 + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
+              //  printf("W[%d %d %d] [%d] Read error %f %d\n", i,j ,k, vps, double_val,100 + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
             }
             else
             {
               read_count++;
               //if (rank == 0)
-              //  fprintf(stderr, "C[%d %d %d] [%d] Read error %f %d\n", i,j ,k, vps, double_val,100+ var + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
+              //  printf("W[%d %d %d] [%d] Read error %f %d\n", i,j ,k, vps, double_val, var /*+ vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i))*/);
             }
           }
         }
-
+        
         else if (strcmp(type_name, FLOAT32) == 0 || strcmp(type_name, FLOAT32_GA) == 0 || strcmp(type_name, FLOAT32_RGB) == 0)
         {
           for (vps = 0; vps < values_per_sample; vps++)
           {
             memcpy(&float_val, data + (index * values_per_sample + vps) * bits_per_sample, bits_per_sample);
-            if (float_val != var + 100 + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)))
+            if (float_val != var + 100 + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)))
             {
               read_error_count++;
-              //if (rank == 100)
-              //fprintf(stderr, "W [%d] [%d %d %d] [%d] Read error %f %d \n", rank, i,j ,k, vps, float_val, var + vps + 100 + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
+              //if (rank == 1)
+              //printf("W [%d] [%d %d %d] [%d] Read error %f %d Diff %f\n", rank, i,j ,k, vps, float_val, var + vps + 100 + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)), (float) (float_val - (var + vps + 100 + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)))));
             }
             else
             {
               read_count++;
-              //if (rank == 100)
-              //  fprintf(stderr, "C [%d %d %d] [%d] Read %f %lld\n", i,j ,k, vps, float_val, 100 + var + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
+              //if (rank == 0)
+              //  printf("C[%d %d %d] [%d] Read %f %lld\n", i,j ,k, vps, data[index * vps[var] + vps], var + vps + ((global_bounds[0] * global_bounds[1]*(local_box_offset[2] + k))+(global_bounds[0]*(local_box_offset[1] + j)) + (local_box_offset[0] + i)));
             }
           }
         }
-
+        
       }
-
-  //fprintf(stderr, "RANK %d CC %d WC %d\n", rank, read_count, read_error_count);
+  
   MPI_Allreduce(&read_count, &total_read_count, 1, MPI_INT, MPI_SUM, MPI_COMM_WORLD);
   MPI_Allreduce(&read_error_count, &total_read_error_count, 1, MPI_INT, MPI_SUM, MPI_COMM_WORLD);
-
+  
   if (rank == 0)
-    fprintf(stderr, "Correct Sample Count %d Incorrect Sample Count %d\n", total_read_count, total_read_error_count);
-
+    printf("Correct Sample Count %d Incorrect Sample Count %d\n", total_read_count, total_read_error_count);
+  
   if(total_read_error_count == 0 && (total_read_count == global_bounds[2]*global_bounds[1]*global_bounds[0]*values_per_sample))
     return 0;
   else
     return 1;
-}
 
-
-//----------------------------------------------------------------
-static void shutdown_mpi()
-{
-#if PIDX_HAVE_MPI
-  MPI_Finalize();
-#endif
 }
