+++ examples/basic-io/idx_write.c
@@ -65,49 +65,19 @@
   #include "utils/PIDX_windows_utils.h"
 #endif
 
-#define MAX_VAR_COUNT 256
-enum { X, Y, Z, NUM_DIMS };
+#include "pidx_examples_utils.h"
 
-static int process_count = 1, rank = 0;
-static unsigned long long rst_box_size[NUM_DIMS];
-static unsigned long long global_box_size[NUM_DIMS];
-static unsigned long long local_box_offset[NUM_DIMS];
-static unsigned long long local_box_size[NUM_DIMS];
-int sub_div[NUM_DIMS];
-static int time_step_count = 1;
-static int variable_count = 1;
-static char output_file_template[512];
-static char var_list[512];
-static unsigned char **data;
-static char output_file_name[512];
-static char var_name[MAX_VAR_COUNT][512];
-static int bpv[MAX_VAR_COUNT];
-static char type_name[MAX_VAR_COUNT][512];
-static int vps[MAX_VAR_COUNT];
+char var_name[MAX_VAR_COUNT][512];
+int bpv[MAX_VAR_COUNT];
+char type_name[MAX_VAR_COUNT][512];
+int vps[MAX_VAR_COUNT];
+PIDX_variable* variable;
+char output_file_template[512];
+char var_list[512];
+char output_file_name[512];
+unsigned char **data;
 
-static PIDX_point global_size, local_offset, local_size, reg_size;
-static PIDX_access p_access;
-static PIDX_metadata_cache cache;
-static PIDX_file file;
-static PIDX_variable* variable;
-
-static void init_mpi(int argc, char **argv);
-static void parse_args(int argc, char **argv);
-static int parse_var_list();
-static int generate_vars();
-static void check_args();
-static void calculate_per_process_offsets();
-static void create_synthetic_simulation_data();
-static void terminate_with_error_msg(const char *format, ...);
-static void terminate();
-static void set_pidx_file(int ts);
-static void set_pidx_variable();
-static void create_pidx_var_point_and_access();
-static void destroy_pidx_var_point_and_access();
-static void destroy_synthetic_simulation_data();
-static void shutdown_mpi();
-static int isNumber(char number[]);
-static char *usage = "Serial Usage: ./idx_write -g 32x32x32 -l 32x32x32 -v 2 -t 4 -f output_idx_file_name\n"
+char *usage = "Serial Usage: ./idx_write -g 32x32x32 -l 32x32x32 -v 2 -t 4 -f output_idx_file_name\n"
                      "Parallel Usage: mpirun -n 8 ./idx_write -g 64x64x64 -l 32x32x32 -v 2 -t 4 -f output_idx_file_name\n"
                      "  -g: global dimensions\n"
                      "  -l: local (per-process) dimensions\n"
@@ -115,6 +85,14 @@ static char *usage = "Serial Usage: ./idx_write -g 32x32x32 -l 32x32x32 -v 2 -t 
                      "  -f: file name template (without .idx)\n"
                      "  -t: number of timesteps\n"
                      "  -v: number of variables (or file containing a list of variables)\n";
+
+static int generate_vars();
+static void parse_args(int argc, char **argv);
+static int parse_var_list();
+static void create_synthetic_simulation_data();
+static void set_pidx_variable(int var);
+static void set_pidx_file(int ts);
+static void destroy_synthetic_simulation_data();
 
 int main(int argc, char **argv)
 {
@@ -138,7 +116,9 @@ int main(int argc, char **argv)
   create_synthetic_simulation_data();
 
   // Create variables
-  create_pidx_var_point_and_access();
+  create_pidx_point_and_access();
+  variable = (PIDX_variable*)malloc(sizeof(*variable) * variable_count);
+  memset(variable, 0, sizeof(*variable) * variable_count);
 
   for (ts = 0; ts < time_step_count; ts++)
   {
@@ -154,55 +134,20 @@ int main(int argc, char **argv)
     PIDX_close(file);
   }
 
-  // Clean up our mess
-  destroy_pidx_var_point_and_access();
+  // Close access and free memory
+  if (PIDX_close_access(p_access) != PIDX_success)
+    terminate_with_error_msg("PIDX_close_access");
+  
+  free(variable);
+  variable = 0;
 
   destroy_synthetic_simulation_data();
-
+  
   shutdown_mpi();
 
   return 0;
 }
 
-//----------------------------------------------------------------
-static void init_mpi(int argc, char **argv)
-{
-  if (MPI_Init(&argc, &argv) != MPI_SUCCESS)
-    terminate_with_error_msg("ERROR: MPI_Init error\n");
-  if (MPI_Comm_size(MPI_COMM_WORLD, &process_count) != MPI_SUCCESS)
-    terminate_with_error_msg("ERROR: MPI_Comm_size error\n");
-  if (MPI_Comm_rank(MPI_COMM_WORLD, &rank) != MPI_SUCCESS)
-    terminate_with_error_msg("ERROR: MPI_Comm_rank error\n");
-}
-
-static int isNumber(char number[])
-{
-    int i = 0;
-
-    //checking for negative numbers
-    if (number[0] == '-')
-        i = 1;
-    for (; number[i] != 0; i++)
-    {
-        //if (number[i] > '9' || number[i] < '0')
-        if (!isdigit(number[i]))
-            return 0;
-    }
-    return 1;
-}
-
-int nextPow2(int v)
-{
-  v--;
-  v |= v >> 1;
-  v |= v >> 2;
-  v |= v >> 4;
-  v |= v >> 8;
-  v |= v >> 16;
-  v++;
-
-  return v;
-}
 
 //----------------------------------------------------------------
 static void parse_args(int argc, char **argv)
@@ -271,10 +216,30 @@ static void parse_args(int argc, char **argv)
   }
 }
 
+static void set_pidx_variable(int var)
+{
+  PIDX_return_code ret = 0;
+  
+  // Set variable name, number of bits, typename
+  ret = PIDX_variable_create(var_name[var], bpv[var] * vps[var], type_name[var], &variable[var]);
+  if (ret != PIDX_success)  terminate_with_error_msg("PIDX_variable_create");
+  
+  // Set the variable offset and size of the local domain,
+  // where the data is in memory (data) and what is its layout in memory (row major)
+  ret = PIDX_variable_write_data_layout(variable[var], local_offset, local_size, data[var], PIDX_row_major);
+  if (ret != PIDX_success)  terminate_with_error_msg("PIDX_variable_write_data_layout");
+  
+  // Tell PIDX that we want to write this variable
+  ret = PIDX_append_and_write_variable(file, variable[var]);
+  if (ret != PIDX_success)  terminate_with_error_msg("PIDX_append_and_write_variable");
+  
+  return;
+}
+
 static int generate_vars(){
-
+  
   int variable_counter = 0;
-
+  
   for(variable_counter = 0; variable_counter < variable_count; variable_counter++){
     int ret;
     int bits_per_sample = 0;
@@ -285,124 +250,15 @@ static int generate_vars(){
     sprintf(temp_name, "var_%d", variable_counter);
     strcpy(var_name[variable_counter], temp_name);
     strcpy(type_name[variable_counter], temp_type_name);
-
+    
     ret = PIDX_values_per_datatype(temp_type_name, &sample_count, &bits_per_sample);
     if (ret != PIDX_success)  return PIDX_err_file;
-
+    
     bpv[variable_counter] = bits_per_sample;
     vps[variable_counter] = sample_count;
   }
-
+  
   return 0;
-}
-
-//----------------------------------------------------------------
-static int parse_var_list()
-{
-  FILE *fp = fopen(var_list, "r");
-  if (fp == NULL)
-  {
-    fprintf(stderr, "Error Opening %s\n", var_list);
-    return PIDX_err_file;
-  }
-
-  int variable_counter = 0, count = 0, len = 0;
-  char *pch1;
-  char line [ 512 ];
-
-  while (fgets(line, sizeof (line), fp) != NULL)
-  {
-    line[strcspn(line, "\r\n")] = 0;
-
-    if (strcmp(line, "(fields)") == 0)
-    {
-      if( fgets(line, sizeof line, fp) == NULL)
-        return PIDX_err_file;
-      line[strcspn(line, "\r\n")] = 0;
-      count = 0;
-      variable_counter = 0;
-
-      while (line[X] != '(')
-      {
-        pch1 = strtok(line, " +");
-        while (pch1 != NULL)
-        {
-          if (count == 0)
-          {
-            char* temp_name = strdup(pch1);
-            strcpy(var_name[variable_counter], temp_name);
-            free(temp_name);
-          }
-
-          if (count == 1)
-          {
-            len = strlen(pch1) - 1;
-            if (pch1[len] == '\n')
-              pch1[len] = 0;
-
-            strcpy(type_name[variable_counter], pch1);
-            int ret;
-            int bits_per_sample = 0;
-            int sample_count = 0;
-            ret = PIDX_values_per_datatype(type_name[variable_counter], &sample_count, &bits_per_sample);
-            if (ret != PIDX_success)  return PIDX_err_file;
-
-            bpv[variable_counter] = bits_per_sample;
-            vps[variable_counter] = sample_count;
-          }
-          count++;
-          pch1 = strtok(NULL, " +");
-        }
-        count = 0;
-
-        if( fgets(line, sizeof line, fp) == NULL)
-          return PIDX_err_file;
-        line[strcspn(line, "\r\n")] = 0;
-        variable_counter++;
-      }
-      variable_count = variable_counter;
-    }
-  }
-  fclose(fp);
-
-  /*
-  int rank = 0;
-  MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-  if (rank == 0)
-  {
-    int v = 0;
-    for(v = 0; v < variable_count; v++)
-      fprintf(stderr, "[%d] -> %s %d %d\n", v, var_name[v], bpv[v], vps[v]);
-  }
-  */
-
-  return PIDX_success;
-}
-
-//----------------------------------------------------------------
-static void check_args()
-{
-  if (global_box_size[X] < local_box_size[X] || global_box_size[Y] < local_box_size[Y] || global_box_size[Z] < local_box_size[Z])
-    terminate_with_error_msg("ERROR: Global box is smaller than local box in one of the dimensions\n");
-
-  // check if the number of processes given by the user is consistent with the actual number of processes needed
-  int brick_count = (int)((global_box_size[X] + local_box_size[X] - 1) / local_box_size[X]) *
-                    (int)((global_box_size[Y] + local_box_size[Y] - 1) / local_box_size[Y]) *
-                    (int)((global_box_size[Z] + local_box_size[Z] - 1) / local_box_size[Z]);
-  if(brick_count != process_count)
-    terminate_with_error_msg("ERROR: Number of sub-blocks (%d) doesn't match number of processes (%d)\n", brick_count, process_count);
-}
-
-//----------------------------------------------------------------
-static void calculate_per_process_offsets()
-{
-  sub_div[X] = (global_box_size[X] / local_box_size[X]);
-  sub_div[Y] = (global_box_size[Y] / local_box_size[Y]);
-  sub_div[Z] = (global_box_size[Z] / local_box_size[Z]);
-  local_box_offset[Z] = (rank / (sub_div[X] * sub_div[Y])) * local_box_size[Z];
-  int slice = rank % (sub_div[X] * sub_div[Y]);
-  local_box_offset[Y] = (slice / sub_div[X]) * local_box_size[Y];
-  local_box_offset[X] = (slice % sub_div[X]) * local_box_size[X];
 }
 
 //----------------------------------------------------------------
@@ -411,13 +267,13 @@ static void create_synthetic_simulation_data()
   int var = 0;
   data = malloc(sizeof(*data) * variable_count);
   memset(data, 0, sizeof(*data) * variable_count);
-
+  
   // Synthetic simulation data
   for(var = 0; var < variable_count; var++)
   {
     unsigned long long i, j, k, val_per_sample = 0;
     data[var] = malloc(sizeof (*(data[var])) * local_box_size[X] * local_box_size[Y] * local_box_size[Z] * (bpv[var]/8) * vps[var]);
-
+    
     unsigned char cvalue = 0;
     short svalue = 0;
     float fvalue = 0;
@@ -428,13 +284,13 @@ static void create_synthetic_simulation_data()
     FILE* fp = fopen("OF_data.txt", "r");
     int o = 0;
     /*
-    for (o = 0; o < 1400; o++)
-    {
-        fscanf(fp, "%f\n", &fvalue);
-        printf("[%d] val = %.16f\n", o, fvalue);
-        memcpy(data[var] + o * sizeof(float), &fvalue, sizeof(float));
-    }
-    */
+     for (o = 0; o < 1400; o++)
+     {
+     fscanf(fp, "%f\n", &fvalue);
+     printf("[%d] val = %.16f\n", o, fvalue);
+     memcpy(data[var] + o * sizeof(float), &fvalue, sizeof(float));
+     }
+     */
     fclose(fp);
 #else
     for (k = 0; k < local_box_size[Z]; k++)
@@ -442,7 +298,7 @@ static void create_synthetic_simulation_data()
         for (i = 0; i < local_box_size[X]; i++)
         {
           unsigned long long index = (unsigned long long) (local_box_size[X] * local_box_size[Y] * k) + (local_box_size[X] * j) + i;
-
+          
           for (val_per_sample = 0; val_per_sample < vps[var]; val_per_sample++)
           {
             if (strcmp(type_name[var], UINT8) == 0 || strcmp(type_name[var], UINT8_GA) == 0 || strcmp(type_name[var], UINT8_RGB) == 0)
@@ -482,48 +338,86 @@ static void create_synthetic_simulation_data()
 }
 
 //----------------------------------------------------------------
-static void terminate()
+static int parse_var_list()
 {
-#if PIDX_HAVE_MPI
-  MPI_Abort(MPI_COMM_WORLD, -1);
-#else
-  exit(-1);
-#endif
-}
-
-//----------------------------------------------------------------
-static void terminate_with_error_msg(const char *format, ...)
-{
-  va_list arg_ptr;
-  va_start(arg_ptr, format);
-  vfprintf(stderr, format, arg_ptr);
-  va_end(arg_ptr);
-  terminate();
-}
-
-//----------------------------------------------------------------
-static void create_pidx_var_point_and_access()
-{
-  // Allocate a PIDX_variable array where we store the information
-  // of all the variables
-  variable = (PIDX_variable*)malloc(sizeof(*variable) * variable_count);
-  memset(variable, 0, sizeof(*variable) * variable_count);
-
-  // Set variables that define the global and local domain information
-  PIDX_set_point(global_size, global_box_size[X], global_box_size[Y], global_box_size[Z]);
-  PIDX_set_point(local_offset, local_box_offset[X], local_box_offset[Y], local_box_offset[Z]);
-  PIDX_set_point(local_size, local_box_size[X], local_box_size[Y], local_box_size[Z]);
-
-  // Set variable that defines the restructuring box size
-  PIDX_set_point(reg_size, rst_box_size[X], rst_box_size[Y], rst_box_size[Z]);
-
-  //  Creating access
-  PIDX_create_access(&p_access);
-  PIDX_set_mpi_access(p_access, MPI_COMM_WORLD);
-
-  PIDX_create_metadata_cache(&cache);
-
-  return;
+  FILE *fp = fopen(var_list, "r");
+  if (fp == NULL)
+  {
+    fprintf(stderr, "Error Opening %s\n", var_list);
+    return PIDX_err_file;
+  }
+  
+  int variable_counter = 0, count = 0, len = 0;
+  char *pch1;
+  char line [ 512 ];
+  
+  while (fgets(line, sizeof (line), fp) != NULL)
+  {
+    line[strcspn(line, "\r\n")] = 0;
+    
+    if (strcmp(line, "(fields)") == 0)
+    {
+      if( fgets(line, sizeof line, fp) == NULL)
+        return PIDX_err_file;
+      line[strcspn(line, "\r\n")] = 0;
+      count = 0;
+      variable_counter = 0;
+      
+      while (line[X] != '(')
+      {
+        pch1 = strtok(line, " +");
+        while (pch1 != NULL)
+        {
+          if (count == 0)
+          {
+            char* temp_name = strdup(pch1);
+            strcpy(var_name[variable_counter], temp_name);
+            free(temp_name);
+          }
+          
+          if (count == 1)
+          {
+            len = strlen(pch1) - 1;
+            if (pch1[len] == '\n')
+              pch1[len] = 0;
+            
+            strcpy(type_name[variable_counter], pch1);
+            int ret;
+            int bits_per_sample = 0;
+            int sample_count = 0;
+            ret = PIDX_values_per_datatype(type_name[variable_counter], &sample_count, &bits_per_sample);
+            if (ret != PIDX_success)  return PIDX_err_file;
+            
+            bpv[variable_counter] = bits_per_sample;
+            vps[variable_counter] = sample_count;
+          }
+          count++;
+          pch1 = strtok(NULL, " +");
+        }
+        count = 0;
+        
+        if( fgets(line, sizeof line, fp) == NULL)
+          return PIDX_err_file;
+        line[strcspn(line, "\r\n")] = 0;
+        variable_counter++;
+      }
+      variable_count = variable_counter;
+    }
+  }
+  fclose(fp);
+  
+  /*
+   int rank = 0;
+   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+   if (rank == 0)
+   {
+   int v = 0;
+   for(v = 0; v < variable_count; v++)
+   fprintf(stderr, "[%d] -> %s %d %d\n", v, var_name[v], bpv[v], vps[v]);
+   }
+   */
+  
+  return PIDX_success;
 }
 
 //----------------------------------------------------------------
@@ -569,43 +463,6 @@ static void set_pidx_file(int ts)
   return;
 }
 
-//----------------------------------------------------------------
-static void set_pidx_variable(int var)
-{
-  PIDX_return_code ret = 0;
-
-  // Set variable name, number of bits, typename
-  ret = PIDX_variable_create(var_name[var], bpv[var] * vps[var], type_name[var], &variable[var]);
-  if (ret != PIDX_success)  terminate_with_error_msg("PIDX_variable_create");
-
-  // Set the variable offset and size of the local domain,
-  // where the data is in memory (data) and what is its layout in memory (row major)
-  ret = PIDX_variable_write_data_layout(variable[var], local_offset, local_size, data[var], PIDX_row_major);
-  if (ret != PIDX_success)  terminate_with_error_msg("PIDX_variable_write_data_layout");
-
-  // Tell PIDX that we want to write this variable
-  ret = PIDX_append_and_write_variable(file, variable[var]);
-  if (ret != PIDX_success)  terminate_with_error_msg("PIDX_append_and_write_variable");
-
-  return;
-}
-
-//----------------------------------------------------------------
-static void destroy_pidx_var_point_and_access()
-{
-  if (PIDX_close_access(p_access) != PIDX_success)
-    terminate_with_error_msg("PIDX_close_access");
-
-  if (PIDX_free_metadata_cache(cache) != PIDX_success)
-    terminate_with_error_msg("PIDX_free_meta_data_cache");
-
-  free(variable);
-  variable = 0;
-
-  return;
-}
-
-//----------------------------------------------------------------
 static void destroy_synthetic_simulation_data()
 {
   int var = 0;
@@ -617,12 +474,3 @@ static void destroy_synthetic_simulation_data()
   free(data);
   data = 0;
 }
-
-//----------------------------------------------------------------
-static void shutdown_mpi()
-{
-#if PIDX_HAVE_MPI
-  MPI_Finalize();
-#endif
-}
-
