+++ examples/basic-io/idx_write_partitioned.c
@@ -35,57 +35,40 @@
   #include "utils/PIDX_windows_utils.h"
 #endif
 
-#define MAX_VAR_COUNT 256
-enum { X, Y, Z, NUM_DIMS };
+#include "pidx_examples_utils.h"
 
-static int process_count = 1, rank = 0;
-static unsigned long long rst_box_size[NUM_DIMS];
-static unsigned long long global_box_size[NUM_DIMS];
-static unsigned long long local_box_offset[NUM_DIMS];
-static unsigned long long local_box_size[NUM_DIMS];
-int sub_div[NUM_DIMS];
-static int time_step_count = 1;
-static int variable_count = 1;
-static char output_file_template[512];
-static char var_list[512];
-static unsigned char **data;
-static char output_file_name[512];
-static char var_name[MAX_VAR_COUNT][512];
-static int bpv[MAX_VAR_COUNT];
-static char type_name[MAX_VAR_COUNT][512];
-static int vps[MAX_VAR_COUNT];
-static int parts_count[3];
+char output_file_template[512];
+char var_list[512];
+unsigned char **data;
+char output_file_name[512];
+char var_name[MAX_VAR_COUNT][512];
+int bpv[MAX_VAR_COUNT];
+char type_name[MAX_VAR_COUNT][512];
+int vps[MAX_VAR_COUNT];
+int parts_count[3];
+PIDX_variable* variable;
 
-static PIDX_point global_size, local_offset, local_size, reg_size;
-static PIDX_access p_access;
-static PIDX_metadata_cache cache;
-static PIDX_file file;
-static PIDX_variable* variable;
+char *usage = "Serial Usage: ./idx_write_partitioned -g 32x32x32 -l 32x32x32 -p 2x1x1 -v 2 -t 4 -f output_idx_file_name\n"
+"Parallel Usage: mpirun -n 8 ./idx_write -g 64x64x64 -l 32x32x32 -p 2x1x1 -v 2 -t 4 -f output_idx_file_name\n"
+"  -g: global dimensions\n"
+"  -l: local (per-process) dimensions\n"
+"  -r: restructured box dimension\n"
+"  -f: file name template (without .idx)\n"
+"  -t: number of timesteps\n"
+"  -v: number of variables (or file containing a list of variables)\n";
 
-static void init_mpi(int argc, char **argv);
 static void parse_args(int argc, char **argv);
 static int parse_var_list();
 static int generate_vars();
-static void check_args();
-static void calculate_per_process_offsets();
 static void create_synthetic_simulation_data();
-static void terminate_with_error_msg(const char *format, ...);
-static void terminate();
+
 static void set_pidx_file(int ts);
 static void set_pidx_variable();
 static void create_pidx_var_point_and_access();
 static void destroy_pidx_var_point_and_access();
 static void destroy_synthetic_simulation_data();
-static void shutdown_mpi();
 static int isNumber(char number[]);
-static char *usage = "Serial Usage: ./idx_write_partitioned -g 32x32x32 -l 32x32x32 -p 2x1x1 -v 2 -t 4 -f output_idx_file_name\n"
-                     "Parallel Usage: mpirun -n 8 ./idx_write -g 64x64x64 -l 32x32x32 -p 2x1x1 -v 2 -t 4 -f output_idx_file_name\n"
-                     "  -g: global dimensions\n"
-                     "  -l: local (per-process) dimensions\n"
-                     "  -r: restructured box dimension\n"
-                     "  -f: file name template (without .idx)\n"
-                     "  -t: number of timesteps\n"
-                     "  -v: number of variables (or file containing a list of variables)\n";
+
 
 int main(int argc, char **argv)
 {
@@ -133,46 +116,6 @@ int main(int argc, char **argv)
   shutdown_mpi();
 
   return 0;
-}
-
-//----------------------------------------------------------------
-static void init_mpi(int argc, char **argv)
-{
-  if (MPI_Init(&argc, &argv) != MPI_SUCCESS)
-    terminate_with_error_msg("ERROR: MPI_Init error\n");
-  if (MPI_Comm_size(MPI_COMM_WORLD, &process_count) != MPI_SUCCESS)
-    terminate_with_error_msg("ERROR: MPI_Comm_size error\n");
-  if (MPI_Comm_rank(MPI_COMM_WORLD, &rank) != MPI_SUCCESS)
-    terminate_with_error_msg("ERROR: MPI_Comm_rank error\n");
-}
-
-static int isNumber(char number[])
-{
-    int i = 0;
-
-    //checking for negative numbers
-    if (number[0] == '-')
-        i = 1;
-    for (; number[i] != 0; i++)
-    {
-        //if (number[i] > '9' || number[i] < '0')
-        if (!isdigit(number[i]))
-            return 0;
-    }
-    return 1;
-}
-
-int nextPow2(int v)
-{
-  v--;
-  v |= v >> 1;
-  v |= v >> 2;
-  v |= v >> 4;
-  v |= v >> 8;
-  v |= v >> 16;
-  v++;
-
-  return v;
 }
 
 //----------------------------------------------------------------
@@ -355,32 +298,6 @@ static int parse_var_list()
 }
 
 //----------------------------------------------------------------
-static void check_args()
-{
-  if (global_box_size[X] < local_box_size[X] || global_box_size[Y] < local_box_size[Y] || global_box_size[Z] < local_box_size[Z])
-    terminate_with_error_msg("ERROR: Global box is smaller than local box in one of the dimensions\n");
-
-  // check if the number of processes given by the user is consistent with the actual number of processes needed
-  int brick_count = (int)((global_box_size[X] + local_box_size[X] - 1) / local_box_size[X]) *
-                    (int)((global_box_size[Y] + local_box_size[Y] - 1) / local_box_size[Y]) *
-                    (int)((global_box_size[Z] + local_box_size[Z] - 1) / local_box_size[Z]);
-  if(brick_count != process_count)
-    terminate_with_error_msg("ERROR: Number of sub-blocks (%d) doesn't match number of processes (%d)\n", brick_count, process_count);
-}
-
-//----------------------------------------------------------------
-static void calculate_per_process_offsets()
-{
-  sub_div[X] = (global_box_size[X] / local_box_size[X]);
-  sub_div[Y] = (global_box_size[Y] / local_box_size[Y]);
-  sub_div[Z] = (global_box_size[Z] / local_box_size[Z]);
-  local_box_offset[Z] = (rank / (sub_div[X] * sub_div[Y])) * local_box_size[Z];
-  int slice = rank % (sub_div[X] * sub_div[Y]);
-  local_box_offset[Y] = (slice / sub_div[X]) * local_box_size[Y];
-  local_box_offset[X] = (slice % sub_div[X]) * local_box_size[X];
-}
-
-//----------------------------------------------------------------
 static void create_synthetic_simulation_data()
 {
   int var = 0;
@@ -454,26 +371,6 @@ static void create_synthetic_simulation_data()
         }
 #endif
   }
-}
-
-//----------------------------------------------------------------
-static void terminate()
-{
-#if PIDX_HAVE_MPI
-  MPI_Abort(MPI_COMM_WORLD, -1);
-#else
-  exit(-1);
-#endif
-}
-
-//----------------------------------------------------------------
-static void terminate_with_error_msg(const char *format, ...)
-{
-  va_list arg_ptr;
-  va_start(arg_ptr, format);
-  vfprintf(stderr, format, arg_ptr);
-  va_end(arg_ptr);
-  terminate();
 }
 
 //----------------------------------------------------------------
@@ -595,12 +492,3 @@ static void destroy_synthetic_simulation_data()
   free(data);
   data = 0;
 }
-
-//----------------------------------------------------------------
-static void shutdown_mpi()
-{
-#if PIDX_HAVE_MPI
-  MPI_Finalize();
-#endif
-}
-
