+++ examples/pidx_examples_utils.h
@@ -1,0 +1,168 @@
+/*****************************************************
+ **  PIDX Parallel I/O Library                      **
+ **  Copyright (c) 2010-2014 University of Utah     **
+ **  Scientific Computing and Imaging Institute     **
+ **  72 S Central Campus Drive, Room 3750           **
+ **  Salt Lake City, UT 84112                       **
+ **                                                 **
+ **  PIDX is licensed under the Creative Commons    **
+ **  Attribution-NonCommercial-NoDerivatives 4.0    **
+ **  International License. See LICENSE.md.         **
+ **                                                 **
+ **  For information about this project see:        **
+ **  http://www.cedmav.com/pidx                     **
+ **  or contact: pascucci@sci.utah.edu              **
+ **  For support: PIDX-support@visus.net            **
+ **                                                 **
+ *****************************************************/
+
+/*
+  PIDX example utils
+*/
+#if !defined _MSC_VER
+#include <unistd.h>
+#endif
+#include <stdarg.h>
+#include <stdint.h>
+#include <ctype.h>
+#include <PIDX.h>
+
+#if defined _MSC_VER
+  #include "utils/PIDX_windows_utils.h"
+#endif
+
+#define MAX_VAR_COUNT 256
+enum { X, Y, Z, NUM_DIMS };
+
+static int process_count = 1, rank = 0;
+static unsigned long long rst_box_size[NUM_DIMS];
+static unsigned long long global_box_size[NUM_DIMS];
+static unsigned long long local_box_offset[NUM_DIMS];
+static unsigned long long local_box_size[NUM_DIMS];
+int sub_div[NUM_DIMS];
+static int time_step_count = 1;
+static int variable_count = 1;
+
+static PIDX_point global_size, local_offset, local_size, reg_size;
+static PIDX_access p_access;
+static PIDX_metadata_cache cache;
+static PIDX_file file;
+
+static void init_mpi(int argc, char **argv);
+static void check_args();
+static void calculate_per_process_offsets();
+static void terminate_with_error_msg(const char *format, ...);
+static void terminate();
+static void destroy_synthetic_simulation_data();
+static void shutdown_mpi();
+static void create_pidx_point_and_access();
+static int isNumber(char number[]);
+
+
+static void create_pidx_point_and_access()
+{
+  PIDX_set_point(global_size, global_box_size[X], global_box_size[Y], global_box_size[Z]);
+  PIDX_set_point(local_offset, local_box_offset[X], local_box_offset[Y], local_box_offset[Z]);
+  PIDX_set_point(local_size, local_box_size[X], local_box_size[Y], local_box_size[Z]);
+  
+  //  Creating access
+  PIDX_create_access(&p_access);
+  PIDX_set_mpi_access(p_access, MPI_COMM_WORLD);
+  
+  return;
+}
+
+//----------------------------------------------------------------
+static void init_mpi(int argc, char **argv)
+{
+  if (MPI_Init(&argc, &argv) != MPI_SUCCESS)
+    terminate_with_error_msg("ERROR: MPI_Init error\n");
+  if (MPI_Comm_size(MPI_COMM_WORLD, &process_count) != MPI_SUCCESS)
+    terminate_with_error_msg("ERROR: MPI_Comm_size error\n");
+  if (MPI_Comm_rank(MPI_COMM_WORLD, &rank) != MPI_SUCCESS)
+    terminate_with_error_msg("ERROR: MPI_Comm_rank error\n");
+}
+
+static int isNumber(char number[])
+{
+    int i = 0;
+
+    //checking for negative numbers
+    if (number[0] == '-')
+        i = 1;
+    for (; number[i] != 0; i++)
+    {
+        //if (number[i] > '9' || number[i] < '0')
+        if (!isdigit(number[i]))
+            return 0;
+    }
+    return 1;
+}
+
+int nextPow2(int v)
+{
+  v--;
+  v |= v >> 1;
+  v |= v >> 2;
+  v |= v >> 4;
+  v |= v >> 8;
+  v |= v >> 16;
+  v++;
+
+  return v;
+}
+
+//----------------------------------------------------------------
+static void check_args()
+{
+  if (global_box_size[X] < local_box_size[X] || global_box_size[Y] < local_box_size[Y] || global_box_size[Z] < local_box_size[Z])
+    terminate_with_error_msg("ERROR: Global box is smaller than local box in one of the dimensions\n");
+
+  // check if the number of processes given by the user is consistent with the actual number of processes needed
+  int brick_count = (int)((global_box_size[X] + local_box_size[X] - 1) / local_box_size[X]) *
+                    (int)((global_box_size[Y] + local_box_size[Y] - 1) / local_box_size[Y]) *
+                    (int)((global_box_size[Z] + local_box_size[Z] - 1) / local_box_size[Z]);
+  if(brick_count != process_count)
+    terminate_with_error_msg("ERROR: Number of sub-blocks (%d) doesn't match number of processes (%d)\n", brick_count, process_count);
+}
+
+//----------------------------------------------------------------
+static void calculate_per_process_offsets()
+{
+  sub_div[X] = (global_box_size[X] / local_box_size[X]);
+  sub_div[Y] = (global_box_size[Y] / local_box_size[Y]);
+  sub_div[Z] = (global_box_size[Z] / local_box_size[Z]);
+  local_box_offset[Z] = (rank / (sub_div[X] * sub_div[Y])) * local_box_size[Z];
+  int slice = rank % (sub_div[X] * sub_div[Y]);
+  local_box_offset[Y] = (slice / sub_div[X]) * local_box_size[Y];
+  local_box_offset[X] = (slice % sub_div[X]) * local_box_size[X];
+}
+
+//----------------------------------------------------------------
+static void terminate()
+{
+#if PIDX_HAVE_MPI
+  MPI_Abort(MPI_COMM_WORLD, -1);
+#else
+  exit(-1);
+#endif
+}
+
+//----------------------------------------------------------------
+static void terminate_with_error_msg(const char *format, ...)
+{
+  va_list arg_ptr;
+  va_start(arg_ptr, format);
+  vfprintf(stderr, format, arg_ptr);
+  va_end(arg_ptr);
+  terminate();
+}
+
+//----------------------------------------------------------------
+static void shutdown_mpi()
+{
+#if PIDX_HAVE_MPI
+  MPI_Finalize();
+#endif
+}
+
